+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -31,7 +31,6 @@ pub enum Event {
     OpenChatBotMessage(Box<OpenChatBotMessage>),
     ReferralCodeAdded(ReferralCodeAdded),
     UserPrincipalUpdated(UpdateUserPrincipalArgs),
-    SecretKeySet(Vec<u8>),
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]