+++ backend/canisters/local_user_index/api/src/updates/c2c_notify_user_index_events.rs
@@ -1,1 +1,12 @@
-pub use crate::c2c_notify_events::*;
+use crate::Event;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub events: Vec<Event>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+}