+++ backend/canisters/user_index/impl/src/updates/c2c_register_bot.rs
@@ -1,5 +1,6 @@
 use crate::{mutate_state, RuntimeState, USER_LIMIT};
 use canister_tracing_macros::trace;
+use event_sink_client::EventBuilder;
 use ic_cdk_macros::update;
 use local_user_index_canister::{Event, UserRegistered};
 use types::{Cycles, UserId};
@@ -60,14 +61,15 @@ fn c2c_register_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
         None,
     );
 
-    state.track_event(
-        "user_registered",
-        now,
-        Some(user_id),
-        crate::UserRegisteredEventPayload {
-            referred: false,
-            is_bot: true,
-        },
+    state.data.event_sink_client.push(
+        EventBuilder::new("user_registered", now)
+            .with_user(user_id.to_string())
+            .with_source(state.env.canister_id().to_string())
+            .with_json_payload(&crate::UserRegisteredEventPayload {
+                referred: false,
+                is_bot: true,
+            })
+            .build(),
     );
 
     Success