+++ backend/external_canisters/sns_archive/api/src/queries/get_transactions.rs
@@ -1,0 +1,4 @@
+use icrc_ledger_types::icrc3::transactions::{GetTransactionsRequest, TransactionRange};
+
+pub type Args = GetTransactionsRequest;
+pub type Response = TransactionRange;