+++ backend/canisters/local_group_index/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.1074](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1074-local_group_index)] - 2024-02-22
+
 ### Added
 
 - VideoCall message + permission + summary/updates ([#5357](https://github.com/open-chat-labs/open-chat/pull/5357))
@@ -14,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Avoid usages of `make_c2c_call` and use macro instead ([#5252](https://github.com/open-chat-labs/open-chat/pull/5252))
 - Propagate video call operators ids for guarding ([#5374](https://github.com/open-chat-labs/open-chat/pull/5374))
 - Use `install_chunked_code` to upgrade Group + Community canisters ([#5412](https://github.com/open-chat-labs/open-chat/pull/5412))
+
+### Fixed
+
+- Fix community canister upgrades ([#5422](https://github.com/open-chat-labs/open-chat/pull/5422))
 
 ## [[2.0.1012](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1012-local_group_index)] - 2024-01-18